feat:  improvements of the paginate on list

+++ src/main/kotlin/com/br/app/cashflowmanagementservice/resource/messaging/entity/MovementStreamPayload.kt
@@ -1,9 +1,7 @@
 package com.br.app.cashflowmanagementservice.resource.messaging.entity
 
-import com.br.app.cashflowmanagementservice.application.enums.TypeMovement
-import com.br.app.cashflowmanagementservice.domain.entities.Movement
+import java.math.BigDecimal
 import java.time.LocalDate
-import java.time.LocalDateTime
 
 data class MovementStreamPayload(
     val id: String?,
@@ -11,5 +9,5 @@ data class MovementStreamPayload(
     val personId: String,
     val date: LocalDate,
     val typeMovement: String,
-    val value: Double
+    val value: BigDecimal
 ): java.io.Serializable
